==> code/MotorStratificationModel/00_data_exploration.Rmd <==
---
title: "description"
author: "Veronica Mandelli"
date: "4/29/2020"
output: html_document
---
In this first script the raw datasets (Downloaded from NDA, and collected at IRCCS-MEDEA) are cleaned and merged.
Standard statistical analysis is run to confirm differences on motor functioning in the 3 groups (TD-ASD-DCD).
```{r, message=FALSE, warning= FALSE, results= 'hide', include=FALSE}
# load libraries

library(easypackages)
# library(fpc)
libraries("here","ggplot2","psych","gplots","ramify","NbClust","cluster","readxl","pracma","MatchIt","grid","amap","gridExtra",
          "apcluster","plsgenomics","clusteval","plot3D","beeswarm")

# set paths
codepath = here("code")
tidydatapath = here("data","tidy")
rowdatapath = here("data")
subjectiddatapath = here ("data","SUBJECTID")
plotpath = here("plots","FINAL")
resultpath = here("results","FINAL")
code_nosharepath = here('code_noshare')

# load functions
source(file.path(code_nosharepath,"cohens_d.R"))

# load color palette
library("MetBrewer")
color = met.brewer("Lakota")
```

### load DBs
and make varibales homogeneous
```{r, message=FALSE, warning= FALSE, results= 'hide', include=FALSE}

# NDAR - NDA dataset
# read the .csv file
DB_NDAR = read.csv(file.path(tidydatapath,"NDAR_cleaned_withNaN.csv"),header = TRUE, stringsAsFactors = FALSE)
DB_NDAR_demo = DB_NDAR #save other varibles for late analysis

# select only needed columns
colnames = c("subjectkey","target","interview_age","sex","mabc2_manualdexterity_std",
             "mabc2_aimingcatching_std","mabc2_balance_std","mabc2_total_std",'title')
DB_NDAR = DB_NDAR[,colnames]

# change variables names to make easier to compare DBs
colnames(DB_NDAR)[which(names(DB_NDAR) == "subjectkey")] <- "subject_id"
colnames(DB_NDAR)[which(names(DB_NDAR) == "interview_age")] <- "age"
colnames(DB_NDAR)[which(names(DB_NDAR) == "title")] <- "study_id"

# transform age in years
DB_NDAR$age=round(DB_NDAR$age/12,2)

# change study id code to make it more readable
for (i in (1:144)) {
    if (DB_NDAR[i, 'study_id'] == "Direct Examination of Imitation-Based Learning in Autism"){
      DB_NDAR[i,'study_id'] = 'NDAR_DEML'
    }
    else if (DB_NDAR[i, 'study_id'] =="A Novel Framework for Impaired Imitation in ASD" ){
      DB_NDAR[i,'study_id'] = 'NDAR_NFII'
    }
    else if (DB_NDAR[i, 'study_id'] == "The neurobiological basis of heterogeneous social and motor deficits in ASD"){
      DB_NDAR[i,'study_id'] = 'NDAR_NBSM'}
}

# exclude people older than 16 years old
DB_NDAR = DB_NDAR[DB_NDAR$age< 17,] # 3 subjects

# BP - IRCCS- MEDEA dataset
# Load the .csv file
DB_BP =  read_excel(file.path(tidydatapath,"BP_abc.xlsx"),
    col_types = c("text", "text","text","numeric", "text", 
        "numeric", "numeric", "numeric", 
        "numeric", "numeric", "numeric", 
        "numeric", "numeric", "numeric", 
        "numeric", "numeric", "numeric", 
        "numeric", "numeric", "numeric", 
        "numeric", "numeric", "numeric", 
        "numeric", "numeric", "numeric" ))

DB_BP_demo = DB_BP #save other varibles for late analysis

# select only needed columns
#colnames = c("subject_id",'age','sex',"group","abc_child_manualdexterity",
#             "abc_child_pointthrow","abc_child_balance","abc_child_total")
colnames = c("ID","diagnosis",'age_years','sex',"MABC2_ManualDexterity_standardscores",
             "MABC2_AimingCatching_standardscores","MABC2_Balance_standardscores","MABC2_Total_StandardScores")
DB_BP = DB_BP[,colnames]


# mask for na in ASD group before clustering
# cols2use  = c("abc_child_manualdexterity","abc_child_pointthrow","abc_child_balance")
cols2use = c("MABC2_ManualDexterity_standardscores","MABC2_AimingCatching_standardscores","MABC2_Balance_standardscores")
mask = rowSums(is.na(DB_BP[,cols2use]))==0
DB_BP = as.data.frame(DB_BP[mask,])

# change variables names to make easier to compare DBs
# colnames(DB_BP)[which(names(DB_BP) == "group")] <- "target"
colnames(DB_BP)[which(names(DB_BP) == "MABC2_ManualDexterity_standardscores")] <- "mabc2_manualdexterity_std"
colnames(DB_BP)[which(names(DB_BP) == "MABC2_AimingCatching_standardscores")] <- "mabc2_aimingcatching_std"
colnames(DB_BP)[which(names(DB_BP) == "MABC2_Balance_standardscores")] <- "mabc2_balance_std"
colnames(DB_BP)[which(names(DB_BP) == "MABC2_Total_StandardScores")] <- "mabc2_total_std"
colnames(DB_BP)[which(names(DB_BP) == "ID")] <- "subject_id"
colnames(DB_BP)[which(names(DB_BP) == "age_years")] <- "age"
colnames(DB_BP)[which(names(DB_BP) == "diagnosis")] <- "target"

# add study id code
DB_BP$study_id = 'BP'

# make sex homogenous  across DBs
for (k in (1:length(DB_BP$sex))) {
   if (DB_BP[k,'sex'] == "0"){
       DB_BP[k,'sex'] = 'M'}
    else {DB_BP[k,'sex'] = "F"}
}

# select only children older than 3 y.o  (ABC-movement not standardized for younger!)
DB_BP = DB_BP[DB_BP$age>=3,] # 3 subject

# save DB_BP_demofor later analysis
savingpath_dataset="/Users/vmandelli/Google Drive/PHD/projects/motorNDAR/data/tidy"
#write.csv(DB_BP_demo,file.path(savingpath_dataset,"DB_BP_demo_renamed.csv"))
```

### modules inferring
The ABC movement battery is dividend into 3 modules selected based on the age of the child. Since the each module has specific tasks here we would like to correct for possible differences due to this difference
module 1: 3-6 years old
module 2: 7-10 years old
module 3: 11-16 years old
```{r,results='hide', warning= FALSE, message=FALSE}

# NDAR
data=DB_NDAR

  for (i in (1:length(data$subject_id))) {
    if (data[i, 'age']<7){
      data[i,'module'] = 1
    }
    else if (data[i, 'age']>=7 && data[i, 'age'] < 11){
      data[i,'module'] = 2
    }
    else if (data[i, 'age']>=11){
      data[i,'module'] = 3}
    
  }

DB_NDAR=data
print(length(subset(DB_NDAR,DB_NDAR$target=='TD')$subject_id))
print(length(subset(DB_NDAR,DB_NDAR$target=='ASD')$subject_id))
print(length(subset(DB_NDAR,DB_NDAR$target=='DCD')$subject_id))


# BP
data=DB_BP


  for (i in (1:length(data$subject_id))) {
    if (data[i, 'age']<7){
      data[i,'module'] = 1
    }
    else if (data[i, 'age']>=7 && data[i, 'age'] < 11){
      data[i,'module'] = 2
    }
    else if (data[i, 'age']>=11){
      data[i,'module'] = 3}
    
  }

DB_BP=data
```

### get some information about age distribution
```{r}
#some descirptive stats

mean_age_bp = mean(DB_BP$age)
std_age_bp = sd(DB_BP$age)
mean_age_nda = mean(DB_NDAR$age)
std_age_nda = sd(DB_NDAR$age)
descr_stat =  data.frame('BP_mean'=mean_age_bp,'BP_sd'=std_age_bp,'NDA_mean'=mean_age_nda,
                         'NDA_sd'=std_age_nda)

#write.csv(descr_stat,file.path(resultpath,"age_mean_sd.csv"))
```

### merge DBs in one big dataset
```{r,results='hide', warning= FALSE, message=FALSE}
# merge the DBs
cols2use = c ('subject_id','module','sex','target','study_id','age','mabc2_manualdexterity_std','mabc2_aimingcatching_std',
              'mabc2_balance_std','mabc2_total_std')

dataset = rbind(DB_BP[,cols2use],DB_NDAR[,cols2use])
# reset row names
rownames(dataset) <- 1:nrow(dataset)

# number of subjects per study by diagnosis
table(dataset$study_id,dataset$target)
print("NDA ASD")
print(18+35+9)
print("NDA TD")
print(20+30+6)
print("TD total")
print(93+20+30+6)
table(dataset$sex,dataset$target)
# total subject per study_id
table(dataset$study_id)
```

### some plots on: MABC2 - variables distribution
```{r}
cols2plot = c("mabc2_total_std","mabc2_manualdexterity_std","mabc2_aimingcatching_std","mabc2_balance_std")
title_list = c("ABC TOTAL","MANUAL DEXTERITY"," AIMING and CATCHING","BALANCE")
dataset$target = factor(dataset$target, level =c("ASD","TD","DCD"))
color = met.brewer("Lakota")
p_list =  list()
for (i in 1:length(cols2plot)){
  var2use = cols2plot[i]
p = ggplot(data =dataset ,aes_string(x = 'target', y = var2use, color='target') )
p = p + geom_jitter() + geom_boxplot(fill = NA, colour = "#000000", outlier.shape = NA)
p = p + scale_colour_manual(values = color)
p = p + guides(colour = FALSE)
p = p + xlab("") + ylab('')
p = p+ scale_y_continuous(name = "",limits=c(1,15), breaks = c(0,5,10,15)) 
#p = p + ylim (1,19)
p = p + ggtitle(title_list[i])
#p = p + theme_bw()
p = p + theme(plot.title = element_text(size = 10, face = "bold",hjust = 0.5))
p = p + theme(axis.text.x = element_text(face="bold", size=10))
p = p + theme(axis.text.y = element_text(face="bold", size=10))
p = p + geom_hline(yintercept=7,color = "black")
p = p + geom_hline(yintercept=5,color = "red")
#p = p + stat_compare_means(method = "t.test", label =  "p.signif", label.x = 1.5)
#p = p + scale_y_continuous( limits=c(0, 72))
print(p)
p_list[[i]]= p}

g = grid.arrange(p_list[[1]],p_list[[2]],p_list[[3]],p_list[[4]],nrow=1)
#ggsave(file.path(plotpath,paste('ABC_asd_td_dcd.png',sep="")), g, width = 10 ,height = 5)
ggsave(file.path(plotpath,paste('ABC_asd_td_dcd.png',sep="")), g, width = 10 ,height = 4)
g

# percentage table
perc_tab =  as.data.frame(matrix(nrow = 3, ncol=3))
diagnosis_group = c("ASD","TD","DCD")
colnames(perc_tab) = diagnosis_group
row_name =  c("norm","at risk","with motor impairment")
rownames(perc_tab) = row_name

for (j in 1:3){
  subset_ = subset(dataset, dataset$target==diagnosis_group[j])
    perc_tab['norm',j] = length(subset_[subset_$mabc2_total_std>=7,'subject_id'])/length(subset_$subject_id)
    perc_tab["at risk",j] = length(subset_[subset_$mabc2_total_std<7 & subset_$mabc2_total_std>5,'subject_id'])/   length(subset_$subject_id)
    perc_tab["with motor impairment",j] = length(subset_[subset_$mabc2_total_std<=5,'subject_id'])/length(subset_$subject_id)
}
perc_tab = round(perc_tab,digit = 2)

#write.csv(perc_tab,file.path(resultpath,"ABC_total_percentage.csv"))
```

### other plot on: age distribution and sex distribution plus modules lines
```{r, warning= FALSE, message=FALSE}

# plot age and sex distribution
data2plot = dataset[dataset$target=='ASD',c('subject_id','age','sex','module')]
data2plot$sex <- factor( data2plot$sex,levels = c("M","F"))

p<-ggplot(data2plot, aes(x=age, fill=sex, color=sex)) +
  geom_histogram(position="identity", alpha=0.4,bins = 30)+
  geom_vline(aes(xintercept=6.5), color="red",linetype="dashed")+
  geom_vline(aes(xintercept=10.5), color="red",linetype="dashed")+
  scale_color_manual(values=c("#99ccff","#ffcccc"))+
  scale_fill_manual(values=c("#99ccff","#ffcccc"))+
  scale_x_continuous( limits=c(3,16),breaks=c(3,4,5,6,7,8,9,10,11,12,13,14,15,16))+
  scale_y_continuous( limits=c(0,20),breaks=c(0,5,10,15,20))+
  labs(title="ASD Sex and Age distribution",x="age (years)", y = "sample size")+
  theme_classic()+coord_fixed(ratio = 0.5)
p
# some numbers 
table(data2plot$module,data2plot$sex)

# saving 
ggsave(file.path(plotpath,"ASD_agesex_distribution.png"),p)
```

### some descirptive stats
```{r}
# some descirptive stats
sample_size = length(data2plot$subject_id)
girl_number = length(data2plot[data2plot$sex=="F",'subject_id'])
boy_number = length(data2plot[data2plot$sex=="M",'subject_id'])
BP_size = length(dataset[dataset$target =="ASD" & dataset$study_id=='BP','subject_id'])
NDAR_DEML_size =length(dataset[dataset$target =="ASD" & dataset$study_id=='NDAR_DEML','subject_id'])
NDAR_NFII_size = length(dataset[dataset$target =="ASD" & dataset$study_id=='NDAR_NFII','subject_id'])
NDAR_NBSM_size = length(dataset[dataset$target =="ASD" & dataset$study_id=='NDAR_NBSM','subject_id'])
mean_age = mean(data2plot$age)
std_age = sd(data2plot$age)
descr_stat =  data.frame('sample_size'=sample_size,'girl_number'=girl_number,'boy_number'=boy_number,
                         'BP_size'=BP_size,'NDAR_DEML_size'=NDAR_DEML_size,'NDAR_NFII_size'=NDAR_NFII_size,
                         'NDAR_NBSM_size'=NDAR_NBSM_size,'mean_age'=mean_age,'std_age'=std_age)
                         
#write.csv(descr_stat,file.path(resultpath,"descr_stat.csv"))                        
```

# extra statistic for distirnguish motor skills in ASD-TD-DCD groups
```{r}
# anova
# Compute the analysis of variance
res.aov <- aov(mabc2_total_std ~ target, data = dataset)
# Summary of the analysis
summary(res.aov)
summary(res.aov)[[1]][["Pr(>F)"]][1]
AIC(res.aov)


#require(tidyverse)
library(rstatix)
library(ggpubr)
library(car)

col2use =  c("mabc2_total_std" )
all_sub_clust = dataset
# prepare table to save statistic
other_feat_mat= as.data.frame(matrix(nrow=2,ncol = 7))
rownames(other_feat_mat) = c("ASD_TD","ASD_DCD")
colnames(other_feat_mat) = c("normality","levene","test_type","t","p","df","CohenD")
list_comarison=c("TD","DCD")
row_list = c("ASD_TD","ASD_DCD")
# t-test requirments
for (i in 1:2 ){
row2use = row_list[i]
comp2use = list_comarison[i]

# t-test requirments
      
      # Compute Shapiro wilk test by goups --> normality
      resp_normality = shapiro.test(all_sub_clust[,col2use])
      #resp_normality
      other_feat_mat[row2use,"normality"] = resp_normality$p.value
      
      # Equality of variance
      
      # Levene's test with one independent variable
      form2use = as.formula(sprintf("%s ~ target",col2use))
      resp_levene = leveneTest(form2use, data = all_sub_clust)
      #resp_levene
      other_feat_mat[row2use,"levene"] = resp_levene$`Pr(>F)`[1]
      
      #if (resp_normality$p.value >0.05 && resp_levene$`Pr(>F)`>0.05 ){
      #  paramentric: t-test
      x = subset(all_sub_clust, all_sub_clust$"target" =="ASD")[,col2use]
      y = subset(all_sub_clust, all_sub_clust$"target" ==comp2use)[,col2use]
      res_t = t.test(x,y)
      #print(res_t)
      other_feat_mat[row2use,"test_type"] = 't-test'
      other_feat_mat[row2use,"t"] = res_t$statistic
      other_feat_mat[row2use,"p"] = res_t$p.value
      other_feat_mat[row2use,"df"] = res_t$parameter
      # }
      # else{# non paramentric: 
      # x = subset(all_sub_clust, all_sub_clust$"target" =="ASD")[,col2use]
      # y = subset(all_sub_clust, all_sub_clust$"target" ==comp2use)[,col2use]
      # res_w = wilcox.test(x, y, alternative = "two.sided")
      # 
      # print(res_t)
      # other_feat_mat[row2use,"test_type"] = 'wilcox'
      # other_feat_mat[row2use,"t"] = res_w$statistic
      # other_feat_mat[row2use,"p"] = res_w$p.value}
      # 
      
      # we also compute the cohens'd to estimate the effect size
      CohenD <- cohens_d(all_sub_clust[all_sub_clust$target=='ASD',col2use],
                       all_sub_clust[all_sub_clust$target==comp2use,col2use], 
                       SIGN=FALSE)
      
      other_feat_mat[row2use,"CohenD"] = CohenD

}
other_feat_mat

#save 
write.csv(other_feat_mat, file.path(resultpath,'t_test_ASDVSTDDCD.csv'))
```

### END
